PRC: got rid of PRC::Proxy
In the old days we routed all virtualisation guest messages to MCP
through the PRC in Dom0. PRC::Proxy was the router process for these
messages. Later PRC::Proxy was only responsible for reading guest
console messages from a FIFO and writing it to NFS. Now this FIFO is a
symlink and we don't need Proxy any more.