Review default usage of hosts/ports - OSRC#1016
There were usage of defaults to the main Tapper host
that might become critical when running unit tests
either with pointing to the central server or
during installation running on localhost.

Remove the default values.

This might trigger the hidden bugs in the
test suites, but that's what we want.