[FEAT] use xl to start Xen guest
PRC now understands a new key to starts guests. This key is called "xen"
and expects a filename. This file is handed to xl with the ending
".xl". If xl does not exist the file is handed to xm with ending
".svm". If the file name ends with any of the above mentioned ending it
is cut of before appending the other ending.