[RFK] testprogram_execute expects hash now
testprogram_execute needs lots of options to know what to do for a
testprogram. These options are taken from a hash ref by the caller
(control_testprogram) and provided as a list. This does not make much
sense. Now the user of the function needs to remember the ordering of
the list and adding a new option to handle in testprogram_execute is
complicated. Therefore, I changed the API of the function to use the
hash we already have.