[FIX] capture: tap handling
This patch fixes a few minor patches around handling or TAP
capturing. The main issue was that the function capture_handler_tap
expects a filename but did not get one. Other issues fixed with this
patch are the missing "use File::Basename", an invalid try to set $\ on
per-handle basis and a mission newline in concatenation of headerlines
and text.