[FIX] don't wait when testprogram is already gone
When the testprogram hits its timeout, we kill it with SIGTERM (15). Now we
give it some time to handle this signal because the timeout might be caused by
testprogram just not knowing its time parameters ;-)
Now if the testprogram handled SIGTERM and therefore does no longer exist,
there is no point in waiting the whole grace period.