[FEAT] improve tapper-client for local test
Tapper must not be a daemon for local testing because that causes issues with
multiple clients running at the same time. For ssh tests it does not need to
be a daemon and actually they could suffer from the same issue as local
tests. Therefore, I removed the App::Daemon code and fork away with
fork(). Use an ENV variable to tell ssh and local tests apart.