Output file names improved
This patch includes the code review sugestion for the file names of
output files. The style is now $filename-001.stdout. This means the
suffix stdout and stderr always stay the last part of the file name and
with a three digit number schema the sorting will be much better.

reviewer: Steffen Schwigon
reviewer: Frank Becker