[TEST]: increase timeout to wait in timeouts
This test tries to read data from PRC and compares it to a given
expected. To make sure the test does not hang up forever if the PRC does
not send data we set up an alarm. If the PRC has to handle timeouts,
this alarm was just a few seconds to short. Therefore, I made it
longer.