[FEAT] allow finding test programs by $PATH
Test programs were always used by absolute paths. When a path was not
absolute, we prepended it with the testprogram path. This prevents calls
to simple programs we can expect to have in $PATH as test program (like
true).
This patch allows both behaviours by no longer prepending the
testprogram path. Instead we put that path into the $PATH environment
variable. That way, testprograms in there will still be used but we also
can use $PATH.