[FIX] set complete program path for testprogram in $PATH
Tapper can now use testprograms from $PATH. This makes it possible to
have everything already installed on the test machine as test program
without having to know the full path.

Instead of adapting all checks to this new feature we set the test
program to the full path of the program in the above mentioned case. Now
all other functions that expect a full path work as expected.