[FIX] capute-tap works with file indices
When the same test program is called multiple times we get the same
outfile name multiple times. PRC prevents overwriting these files by
appending a number to the later output files (-001, -002, ...). When the
test suite does not want to send the TAP report to the Tapper reports
framework it can be called with "capture: tap" or "capture:
tap-stderr". In this case, PRC will send the report.
This patch makes sure these two features work together.