[FIX] sleep for correct grace period
sleep($time) is not guaranteed to sleep exactly the given $time. It only
guarantees that it won't sleep less. Coordination between MCP and PRC
depends on knowing how long the PRC waits for the testprogram to clean
up. Therefore sixty times sleep(1) is not close enough to sleeping 60
seconds.
This fix now checks the grace period for testprogram cleanup based on
time(). Even though, this still isn't guaranteed to lead to exactly 60
seconds of sleep time, we get close enough to it for correct
coordination with MCP.