test_type 'minion' uses different output dir
Because we now have two job queues, Tapper MCP and
Minion, the timing of handling file uploads and cleanup
is a bit difficult to synchronize. In particular, when
running on the same machine, the MCP might already clean
up files right after it sent the stop signal to Minion.

There it could take some time until it actually handles
the signal and wants to upload files.

So we now simply use directories not used by the MCP to
avoid this problem.